requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.9
pydantic==1.10.13
python-dotenv==1.0.1
pytube==15.0.0
youtube-transcript-api==1.0.2
youtube-search